requirements.txt
# Core Django
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
sqlparse==0.5.3
tzdata==2025.2

# Image processing (for product images)
pillow==11.3.0

# HTTP requests & utilities
requests==2.32.5
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0

# Authentication & Security
PyJWT==2.10.1

# Configuration
python-dotenv==1.1.1

# Email service - Resend for reliable email delivery
resend==2.6.0

# AI Service - OpenAI for intelligent pricing
openai==1.55.0

# Template engine (if using Jinja2 templates)
Jinja2==3.1.4
MarkupSafe==2.1.5

# Utilities
packaging==25.0
setuptools==70.2.0
typing_extensions==4.12.2
colorama==0.4.6
regex==2025.9.1
PyYAML==6.0.2